# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-throughput event ingestion and analytics service backed by ClickHouse.\nSupports single and bulk event ingestion with async batched writes,\nin-memory deduplication, and aggregated metrics queries.",
        "title": "Events",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/events": {
            "post": {
                "description": "Validates and enqueues a single event for async persistence to ClickHouse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Ingest a single event",
                "parameters": [
                    {
                        "description": "Event payload",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.EventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Duplicate event (idempotent)",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.EventResponse"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/bulk": {
            "post": {
                "description": "Validates and enqueues multiple events for async persistence",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Ingest events in bulk",
                "parameters": [
                    {
                        "description": "Bulk event payload",
                        "name": "events",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.BulkEventRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.BulkEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health/live": {
            "get": {
                "tags": [
                    "health"
                ],
                "summary": "Liveness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health/ready": {
            "get": {
                "description": "Checks ClickHouse connectivity and returns pipeline stats",
                "tags": [
                    "health"
                ],
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "description": "Returns total count, unique users, and optional breakdown by channel or time bucket",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Query aggregated event metrics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event name filter",
                        "name": "event_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Start timestamp (unix seconds)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "End timestamp (unix seconds)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Aggregation dimension: channel, hourly, daily",
                        "name": "group_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.MetricsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/events_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "events_internal_dto.BulkEventRequest": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/events_internal_dto.EventRequest"
                    }
                }
            }
        },
        "events_internal_dto.BulkEventResponse": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "integer",
                    "example": 95
                },
                "duplicates": {
                    "type": "integer",
                    "example": 3
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rejected": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "events_internal_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string",
                    "example": "validation failed"
                }
            }
        },
        "events_internal_dto.EventRequest": {
            "type": "object",
            "properties": {
                "campaign_id": {
                    "type": "string",
                    "example": "cmp_987"
                },
                "channel": {
                    "type": "string",
                    "example": "web"
                },
                "event_name": {
                    "type": "string",
                    "example": "product_view"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "electronics",
                        "homepage"
                    ]
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1723475612
                },
                "user_id": {
                    "type": "string",
                    "example": "user_123"
                }
            }
        },
        "events_internal_dto.EventResponse": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4"
                },
                "status": {
                    "type": "string",
                    "example": "accepted"
                }
            }
        },
        "events_internal_dto.MetricBreakdown": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string",
                    "example": "web"
                },
                "total_count": {
                    "type": "integer",
                    "example": 9200
                },
                "unique_users": {
                    "type": "integer",
                    "example": 5100
                }
            }
        },
        "events_internal_dto.MetricsResponse": {
            "type": "object",
            "properties": {
                "breakdown": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/events_internal_dto.MetricBreakdown"
                    }
                },
                "event_name": {
                    "type": "string",
                    "example": "product_view"
                },
                "from": {
                    "type": "integer",
                    "example": 1723400000
                },
                "to": {
                    "type": "integer",
                    "example": 1723500000
                },
                "total_count": {
                    "type": "integer",
                    "example": 15432
                },
                "unique_users": {
                    "type": "integer",
                    "example": 8721
                }
            }
        }
    }
}